refactor: flatten core/domain subfolders and finalize folder structure
- Flatten core/domain/instrument/ → classifier.js directly in domain/
- Flatten core/domain/parser/pdf/ → files moved into parser/
- Update all imports referencing the moved files
- Update CLAUDE.md to reflect final folder layout

Closes #111
