+++ src/core/domain/parser/PdfTableExtractor.js
