+++ src/core/domain/parser/PdfTradeConfirmationParser.js
